Widen two narrow taxi-path loop counters
Both loops compared a narrow counter against std::vector::size(). The
comparison itself is value-correct — the counter promotes — but the
counter cannot represent every index the container may hold, so a
container larger than the counter's range wraps it back to zero and the
loop never terminates.

BuildForwardReservationPlan iterated path->tiles with a uint16_t. Taxi
paths are nowhere near 65535 tiles, so this was not reachable.

FindTaxiSegmentIndex iterated path->segments with a uint8_t, and the
margin there is thinner: segments.size() is bounded only by tiles.size(),
not by anything structural that keeps it under 256.

Both counters are now size_t. FindTaxiSegmentIndex keeps its uint8_t
return type — taxi_current_segment is a uint8_t field — so the index is
narrowed explicitly at the return.

Reported by CodeQL as cpp/comparison-with-wider-type.

Verified: openttd_test "ModularAirport*" 6183 assertions in 42 test
cases; regression mass7-inair 9251, helis2 13816, T5j2 6309, all
unchanged from the previous baseline.

+++ src/modular_airport_cmd.cpp
@@ -1237,7 +1237,7 @@ static ForwardPlanStatus BuildForwardReservationPlan(const Aircraft *v, const St
 		SortAndUniqueTiles(operation_resource_tiles);
 	}
 
-	for (uint16_t i = start_index; i < path->tiles.size(); ++i) {
+	for (size_t i = start_index; i < path->tiles.size(); ++i) {
 		const TileIndex tile = path->tiles[i];
 		const bool on_operation_runway = ContainsSortedTile(operation_resource_tiles, tile);
 		if (on_operation_runway) {
@@ -2941,9 +2941,9 @@ void ClearTaxiPathState(Aircraft *v, TileIndex keep_tile)
 uint8_t FindTaxiSegmentIndex(const TaxiPath *path, uint16_t tile_index)
 {
 	if (path == nullptr) return 0;
-	for (uint8_t i = 0; i < path->segments.size(); ++i) {
+	for (size_t i = 0; i < path->segments.size(); ++i) {
 		const TaxiSegment &seg = path->segments[i];
-		if (tile_index >= seg.start_index && tile_index <= seg.end_index) return i;
+		if (tile_index >= seg.start_index && tile_index <= seg.end_index) return static_cast<uint8_t>(i);
 	}
 	return static_cast<uint8_t>(path->segments.size());
 }
